Renamed and relocated filesthroughout the app to make it more manageable and easier to maintain
