+++ src/gui/stack_compo/home_widget.py
