+++ src/gui/stack_compo/view_sessions_widget.py
