+++ src/logic/data_classes/new_session_data.py
