+++ src/logic/results_session/calculate_results_session.py
